+++ app/Services/Rest.php
@@ -87,6 +87,7 @@ class Rest extends Directory_Listing_Base implements Service {
 		// use our own hooks to allow import of REST API media files.
 		add_filter( 'eml_mime_type_for_multiple_files', array( $this, 'allow_json_response' ), 10, 3 );
 		add_filter( 'eml_filter_url_response', array( $this, 'get_rest_api_files' ), 10, 3 );
+		add_filter( 'efml_service_rest_hide_file', array( $this, 'prevent_not_allowed_files' ), 10, 2 );
 	}
 
 	/**
@@ -191,6 +192,25 @@ class Rest extends Directory_Listing_Base implements Service {
 				foreach ( $files as $file ) {
 					// bail if source_url is not set.
 					if ( ! isset( $file['source_url'] ) ) {
+						continue;
+					}
+
+					// get the mime type.
+					$mime_type = $file['mime_type'];
+
+					$false = false;
+					/**
+					 * Filter whether given local file should be hidden.
+					 *
+					 * @since 5.0.0 Available since 5.0.0.
+					 *
+					 * @param bool $false True if it should be hidden.
+					 * @param string $mime_type The used mime type.
+					 * @param string $directory The requested directory.
+					 *
+					 * @noinspection PhpConditionAlreadyCheckedInspection
+					 */
+					if ( apply_filters( Init::get_instance()->get_prefix() . '_service_rest_hide_file', $false, $mime_type, $directory ) ) {
 						continue;
 					}
 
@@ -314,6 +334,20 @@ class Rest extends Directory_Listing_Base implements Service {
 	 * @return array<string,mixed>
 	 */
 	public function get_translations( array $translations ): array {
+		// get the method.
+		$method = filter_input( INPUT_GET, 'method', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+
+		// bail if no method is called.
+		if( is_null( $method ) ) {
+			return $translations;
+		}
+
+		// bail if called method is not ours.
+		if( 'rest' !== $method ) {
+			return $translations;
+		}
+
+		// add our custom translation.
 		$translations['form_file'] = array(
 			'title'       => __( 'Enter the WordPress-URL', 'external-files-in-media-library' ),
 			'description' => __( 'Enter the URL of the WordPress project from which you want to integrate media files into your project via REST API.', 'external-files-in-media-library' ),
@@ -324,6 +358,8 @@ class Rest extends Directory_Listing_Base implements Service {
 				'label' => __( 'Use this URL', 'external-files-in-media-library' ),
 			),
 		);
+
+		// return the resulting translations.
 		return $translations;
 	}
 
@@ -454,27 +490,39 @@ class Rest extends Directory_Listing_Base implements Service {
 						continue;
 					}
 
-					// get content type of this file.
-					$mime_type = wp_check_filetype( $file['source_url'] );
-
-					// bail if file is not allowed.
-					if ( empty( $mime_type['type'] ) ) {
-						continue;
-					}
-
 					// check whether to save this file local or let it extern.
-					$local = $http_obj->url_should_be_saved_local( $url, $mime_type['type'] );
+					$local = $http_obj->url_should_be_saved_local( $url, $file['mime_type'] );
 
-					// add this file to the list.
-					$file_list[] = array(
+					// collect the file data.
+					$entry = array(
 						'title'     => basename( $file['source_url'] ),
 						'filesize'  => isset( $file['media_details']['filesize'] ) ? absint( $file['media_details']['filesize'] ) : 0,
-						'mime-type' => $mime_type['type'],
+						'mime-type' => $file['mime_type'],
 						'local'     => $local,
 						'url'       => $file['source_url'],
 						'tmp-file'  => $local ? $http_obj->get_temp_file( $file['source_url'], Helper::get_wp_filesystem() ) : '',
 						'last-modified' => absint( strtotime( $file['modified'] ) )
 					);
+
+					$response_headers = array();
+					/**
+					 * Filter the data of a single file during import.
+					 *
+					 * @since        1.1.0 Available since 1.1.0
+					 *
+					 * @param array<string,mixed>  $results List of detected file settings.
+					 * @param string $url     The requested external URL.
+					 * @param array<string,mixed> $response_headers The response header.
+					 */
+					 $entry = apply_filters( 'eml_external_file_infos', $entry, $url, $response_headers );
+
+					 // bail if entry is empty.
+					if( empty( $entry ) ) {
+						continue;
+					}
+
+					// add entry to the list.
+					$file_list[]  = $entry;
 				}
 			} catch ( JsonException $e ) {
 				continue;
@@ -483,5 +531,23 @@ class Rest extends Directory_Listing_Base implements Service {
 
 		// return the resulting list of files.
 		return $file_list;
+	}
+
+	/**
+	 * Prevent visibility of not allowed mime types.
+	 *
+	 * @param bool   $result The result - should be true to prevent the usage.
+	 * @param string $mime_type The used mime type.
+	 *
+	 * @return bool
+	 */
+	public function prevent_not_allowed_files( bool $result, string $mime_type ): bool {
+		// bail if setting is disabled.
+		if( 1 !== absint( get_option( 'eml_directory_listing_hide_not_supported_file_types' ) ) ) {
+			return $result;
+		}
+
+		// return whether this file type is allowed (false) or not (true).
+		return ! in_array( $mime_type, Helper::get_allowed_mime_types(), true );
 	}
 }