+++ changelog.md
@@ -6,9 +6,9 @@
 
 - Added custom REST API support for handling of external URLs from other WordPress-projects in your media library
 - Added option to delete synchronized files of single directory archive with one click
-- Added file type specific icons in directory listings
-- Added option for use the date of external files in add-dialog
+- Added option to use the date of external files in add-dialog
 - Added option to really import files in media library (this disabled all external files functions for these files)
+- Added file type specific icons in directory listings
 - Added Taskfile as third way to build plugin release
 - Added check for PHP strict usage on every release with PHPStan
 
@@ -26,6 +26,7 @@
 - Active folder in directory listing is now marked
 - Optimized styling of directory listings
 - Standardize the usage of timestamp as last-modified date for each service and protocol
+- Optimized check if a mime type is allowed in directory listing and during import
 - Multiple new hooks
 - Extended documentation in GitHub for all services we provide